refactor(models): move HTTP helpers from driver_http.go into base_model.go
Per review feedback ("put these in base_model.go"), consolidate the shared
driver utilities into base_model.go: NewDriverHTTPClient, PostJSONRequest, and
ReadErrorBody now live alongside ParseSSEStream. driver_http.go is removed.
