+++ internal/entity/models/base_model.go
@@ -18,11 +18,14 @@ package models
 
 import (
 	"bufio"
+	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"strings"
+	"time"
 )
 
 type BaseModel struct {
@@ -132,4 +135,57 @@ func ParseListModel(modelList ModelList) []ListModelResponse {
 		models = append(models, modelResponse)
 	}
 	return models
+}
+
+// NewDriverHTTPClient returns an *http.Client with the standard connection-pool
+// settings used by every model driver. It clones http.DefaultTransport when
+// possible so proxy settings (HTTP_PROXY / HTTPS_PROXY / NO_PROXY) and any
+// process-wide transport customisation are inherited automatically.
+//
+// Drivers should call this once in their constructor and store the result in
+// baseModel.httpClient; do not create a bare &http.Client{} inline.
+func NewDriverHTTPClient() *http.Client {
+	var t *http.Transport
+	if dt, ok := http.DefaultTransport.(*http.Transport); ok {
+		t = dt.Clone()
+	} else {
+		t = &http.Transport{Proxy: http.ProxyFromEnvironment}
+	}
+	t.MaxIdleConns = 100
+	t.MaxIdleConnsPerHost = 10
+	t.IdleConnTimeout = 90 * time.Second
+	t.DisableCompression = false
+	t.ResponseHeaderTimeout = 60 * time.Second
+	return &http.Client{Transport: t}
+}
+
+// PostJSONRequest marshals body to JSON, creates a POST request to url using
+// client, and sets the Content-Type header to application/json.  If auth is
+// non-empty it is set as the Authorization header verbatim (include the scheme,
+// e.g. "Bearer sk-...").
+//
+// The caller is responsible for closing resp.Body.
+func PostJSONRequest(ctx context.Context, client *http.Client, url, auth string, body map[string]interface{}) (*http.Response, error) {
+	data, err := json.Marshal(body)
+	if err != nil {
+		return nil, fmt.Errorf("failed to marshal request: %w", err)
+	}
+	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, bytes.NewBuffer(data))
+	if err != nil {
+		return nil, fmt.Errorf("failed to create request: %w", err)
+	}
+	req.Header.Set("Content-Type", "application/json")
+	if auth != "" {
+		req.Header.Set("Authorization", auth)
+	}
+	return client.Do(req)
+}
+
+// ReadErrorBody reads all bytes from r and returns them as a string suitable
+// for embedding in an error message. It never returns an error of its own;
+// if reading fails it returns an empty string so the caller's error path stays
+// simple.
+func ReadErrorBody(r io.Reader) string {
+	b, _ := io.ReadAll(r)
+	return string(b)
 }
